Fix bug with Google gmail and remarks bug
Google gmail has issue with emails with same title and name being sent to trash from mobile. It gets referenced to the wrong email.
Added a new feature to just scan all emails and ignore if it was a repeated email. so deleting multiple reconfirmed email is no longer necessary.
There is a bug when customer enters remarks with special characters like '<' and '|' added control to mitigate these.